# _tmp/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SakuraFilter API",
    "description": "工业/汽车滤清器产品管理平台 API\r\n\r\n## 模块说明\r\n- **认证**: JWT 登录/刷新/登出/用户管理\r\n- **公开搜索**: 前台产品搜索/详情 (无需认证)\r\n- **后台产品管理**: CRUD (需 admin 角色)\r\n- **字典管理**: 8 类字典 CRUD (需认证)\r\n- **ETL**: 数据导入/状态/进度 (X-Admin-Token)\r\n- **运维**: 健康检查/指标/性能告警\r\n\r\n## 认证方式\r\n1. **JWT Bearer** (推荐): 登录 /api/auth/login 获取 token, 放入 Authorization: Bearer {token}\r\n2. **X-Admin-Token** (ETL/CI 备用): 放入 X-Admin-Token: {token}",
    "contact": {
      "name": "SakuraFilter Team",
      "email": "admin@sakurafilter.dev"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/products/{oem}": {
      "get": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "产品详情 (按 OEM 精确/规范化查询, 含 cross-references + machine-applications)",
        "operationId": "GetProductByOem",
        "parameters": [
          {
            "name": "oem",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/products": {
      "post": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "后台创建产品 (含 cross-references + machine-applications 嵌套创建)",
        "operationId": "AdminCreateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductFormDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "后台产品列表 (支持搜索/筛选/排序/分页, 含 published/discontinued 状态)",
        "operationId": "AdminListProducts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDiscontinued",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/products/{id}": {
      "get": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "后台产品详情 (含全部字段, 不排除下架)",
        "operationId": "AdminGetProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "后台更新产品 (xmin 乐观锁, 409 冲突)",
        "operationId": "AdminUpdateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductFormDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "SakuraFilter.Api"
        ],
        "summary": "后台软删除产品 (is_discontinued=true, 保留历史)",
        "operationId": "AdminDeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MachineAppInput": {
        "type": "object",
        "properties": {
          "machineBrand": {
            "type": "string",
            "nullable": true
          },
          "machineModel": {
            "type": "string",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "nullable": true
          },
          "engineBrand": {
            "type": "string",
            "nullable": true
          },
          "engineType": {
            "type": "string",
            "nullable": true
          },
          "engineEnergy": {
            "type": "string",
            "nullable": true
          },
          "productionDateStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "productionDateEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "power": {
            "type": "string",
            "nullable": true
          },
          "serialNumberFrom": {
            "type": "string",
            "nullable": true
          },
          "serialNumberTo": {
            "type": "string",
            "nullable": true
          },
          "carBodyType": {
            "type": "string",
            "nullable": true
          },
          "series": {
            "type": "string",
            "nullable": true
          },
          "co2EmissionStandard": {
            "type": "string",
            "nullable": true
          },
          "transmissionType": {
            "type": "string",
            "nullable": true
          },
          "engineDisplacement": {
            "type": "string",
            "nullable": true
          },
          "numberOfCylinders": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gvwr": {
            "type": "string",
            "nullable": true
          },
          "tonnage": {
            "type": "string",
            "nullable": true
          },
          "geographicArea": {
            "type": "string",
            "nullable": true
          },
          "chassisType": {
            "type": "string",
            "nullable": true
          },
          "engineModel": {
            "type": "string",
            "nullable": true
          },
          "cabinType": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "string",
            "nullable": true
          },
          "engineSerialNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProductFormDto": {
        "type": "object",
        "properties": {
          "oem2": {
            "type": "string",
            "nullable": true
          },
          "productName1": {
            "type": "string",
            "nullable": true
          },
          "productName2": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "mr1": {
            "type": "string",
            "nullable": true
          },
          "isPublished": {
            "type": "boolean"
          },
          "remark": {
            "type": "string",
            "nullable": true
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "d1Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "d2Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "d3Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "d4Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "h1Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "h2Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "h3Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "h4Mm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "d7Thread": {
            "type": "string",
            "nullable": true
          },
          "d8Thread": {
            "type": "string",
            "nullable": true
          },
          "noCheckValves": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "noBypassValves": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "noCheckValvesRaw": {
            "type": "string",
            "nullable": true
          },
          "noBypassValvesRaw": {
            "type": "string",
            "nullable": true
          },
          "media": {
            "type": "string",
            "nullable": true
          },
          "mediaModel": {
            "type": "string",
            "nullable": true
          },
          "bypassValveLr": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bypassValveHr": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "efficiency1": {
            "type": "string",
            "nullable": true
          },
          "efficiency2": {
            "type": "string",
            "nullable": true
          },
          "bypassPressure": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "collapsePressureBar": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "bypassValveLrRaw": {
            "type": "string",
            "nullable": true
          },
          "bypassValveHrRaw": {
            "type": "string",
            "nullable": true
          },
          "bypassPressureRaw": {
            "type": "string",
            "nullable": true
          },
          "collapsePressureBarRaw": {
            "type": "string",
            "nullable": true
          },
          "sealingMaterial": {
            "type": "string",
            "nullable": true
          },
          "tempRange": {
            "type": "string",
            "nullable": true
          },
          "qtyPerCarton": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weightKgs": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cartonLengthMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cartonWidthMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cartonHeightMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "masterBoxQty": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "masterBoxWeightKgs": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "masterBoxLengthMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "masterBoxWidthMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "masterBoxHeightMm": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "crossReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/XrefInput"
            },
            "nullable": true
          },
          "machineApplications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MachineAppInput"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "XrefInput": {
        "type": "object",
        "properties": {
          "productName1": {
            "type": "string",
            "nullable": true
          },
          "oemBrand": {
            "type": "string",
            "nullable": true
          },
          "oemNo3": {
            "type": "string",
            "nullable": true
          },
          "oem2": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "machineType": {
            "type": "string",
            "nullable": true
          },
          "isPublished": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Admin-Token": {
        "type": "apiKey",
        "description": "Day 8.4: dev 静态 token, 从 appsettings.json 的 Auth:DevStaticToken 读",
        "name": "X-Admin-Token",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token. 格式: Bearer {token} (登录 /api/auth/login 获取)",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "X-Admin-Token": [ ]
    },
    {
      "Bearer": [ ]
    }
  ],
...
}